Fix terraform bugs with research env and private gateway (#85)
Fixes a few edge cases in the Terraform deploy module:
 - tfstate storage accounts must be alphanumeric and lowercase 
- WAF must not be defined _at all_ in the template when disabling it,
not just set `enabled=false`
- V2 App Gateways _must_ use a public IP, unless the
`EnableApplicationGatewayNetworkIsolation` preview feature is enabled in
a subscription. This option does not appear to be available in GovCloud.
A workaround for private deployments with V2 gateways is to set up a
public IP but not attach any listeners to it (so the app is not actually
exposed to inbound traffic; outbound traffic can be limited with other
settings). See
https://learn.microsoft.com/en-us/azure/application-gateway/application-gateway-private-deployment?tabs=portal
for details.
- The OpenAI custom subdomain must use lowercase, or it will be
re-created every deployment (since it coerces values to lowercase upon
creation, but terraform itself is case sensitive).
- Fix the cryptic "volume does not defined" bug by passing both
`secrets` and `data` volumes to the `PATCH` request when updating the
container app resource, and not assuming the new volumes in this request
will be merged with old ones. Unclear how this worked previously.
 - Update documentation