Support for JSON output format (#90)
Support for `outputFormat=JSON` in the input schema to return redaction
results as an object with `original` and `redacted` fields of plain
text.

Compare `outputFormat=TEXT` which shows the rendered redaction result
with some additional formatting of titles and disclaimers. The `JSON`
format instead contains input and output content.

Example:

```json
{
  "document": {
    "documentId": "foo",
    "attachmentType": "JSON",
    "content": {
      "original": "John was ...",
      "redacted": "[Person 1] was ..."
    }
  }
}
```

The input and output `Document` types are now no longer symmetrical,
since it doesn't make sense to input this schema in the request. The
`Document` schema type has been split into `InputDocument` and
`OutputDocument` accordingly.