Improve status messages (#93)
- Store all task IDs in the chain when it's created, since a quirk of
celery is that they can't be recovered later
- Inspect all task states in the chain instead of just the final one
(which is the default) to get a more informative status
- Add special handling for the RETRY state so that we can save/show the
exception that inspired the retry
- Add some more detail to the retry status response that includes
details about how many times the job has been attempted, as well as a
timestamp of the last failed attempt. This should hopefully defeat the
illusion that jobs are "stuck in PENDING or QUEUED states," which is a
report we've gotten a lot. (In reality nothing is stuck, just that we
retry over the course of several minutes and don't surface the
underlying error info.
- Shorten the retry wait timeout to 30s, with backoff. This is down from
3 minutes with backoff, which also contributes to the illusion that jobs
are "stuck." Now an error will be returned within around 2 (and less
than 5) minutes.
 - Fix a bug where errors weren't always shown at the end of processing.
 - A bit of refactoring to support the above