Fix default status code in generated code to reflect schema (#94)
Schema described our `POST` endpoints as returning `HTTP 201` in the
normal state, but the handlers were all returning 200. This was due to a
shortcoming in the generated code which did not set the default response
code. This is now fixed both in the underlying code generator and in the
template used to generate from the schema.