Fix status check transient errors (#96)
Status check on Azure returns "error" occasionally when there is an
expired connection to a worker that's reused for `ping`. This seems to
be a known issue and doesn't seem to actually impact the usability of
the queue.

The status check will now try to check status up to two times and report
"healthy" with a warning if we hit this issue on the first request.

There may be a better fix, such as purging the bad connections before
trying to test. If the "warning" gets too obnoxious at some point in the
future we can fix this.

Also bumps the version of the redis python client to pick up other bug
fixes (perhaps even one that will mitigate this issue).