Private networking for Azure Monitor (#99)
Fixes an issue with metrics collection. PrivateLink was not configured
for the Azure Monitor service, and the outbound firewall was also not
configured to let the metrics collector reach the Azure Monitor service.
This resulted in metric data failing to reach the Monitor service, so no
data on app performance has been collected since the outbound firewall
restrictions were put in place. Further, the fail/retry disk cache that
the local metrics collector uses was filled and spewing out lots of
warnings and cluttering the logs, ironically making debugging even
harder.

Fixed by setting up private networking for the Monitor. This is a little
more complicated than other services. It requires a private endpoint
deployment in a new subnet and additionally setting up a private link
scoped service for the monitor.

Luckily, Azure appears to adapt to the private link networking
automatically and not require changing the connection string used to
connect to the Monitor service.

This PR adds an additional security restriction that metrics _must_ be
ingested over PrivateLink.

---

All partners should apply this patch ASAP in case we need to debug
something. Having the metrics properly logged and the console logs
cleaner will help tremendously.

Instructions for upgrading are:

1. Connect to cloud shell where you previously ran deployments
2. Navigate to the terraform directory, e.g. `cd
blind-charging-api/terraform`
3. Run `git pull` to fetch the latest updates
4. Run an upgrade, e.g `terraform apply
-var-file="../../<my-site>.tfvars"` (replacing `<my-site>` with the
appropriate file name).
5. Review the proposed changes -- should be some new networking items
related to monitor.
6. Type "yes" and wait for the updates to complete.