Set recommended Redis capacity SKU to Standard-C4 (#103)
The current default of Standard-C3 is not significantly cheaper than
Standard-C4, while C4 offers 2x memory which will give even larger
deployments quite a large amount of headroom. This will also make the
idea of extending the retention period on the results store less
concerning.

The tradeoff with C4 vs. C3 (i.e., the reason it is similarly priced) is
that it offers reduced network capacity. This is not expected to be a
bottleneck in our application, so it's an easy tradeoff to make.