Drop `latest` channel (#125)
Follow up to #123 and #124 to fully deprecate `latest` channel. This
means we will *not* push to `latest` anymore. Untested updates will go
to `unstable` and official version bumps will go to `stable`.

Doing this now has the convenient property of preventing a regression
from unintended image updates without a corresponding terraform update
(as noted in #123).