Improvements to case-level metadata (#126)
A few improvements to handling of inferred data from document redaction
runs.

 - Use the new runtime parameters expected by the new version of `bc2`
- Track a new, less-structured, map of placeholders we've inferred and
join them into the mask info before running redaction
- Save inferred id->mask and name->mask maps in redis with other case
data, for re-use in related documents

This should all make redactions for documents within the same case more
consistent. In addition, we should do a better job conforming to the
`subjects` map sent with the request as well as producing the id->mask
map in the output as promised.